+++ noderedpy/__init__.py
@@ -1,19 +1,17 @@
 # -*- coding: utf-8 -*-
-from ._nodered import RED, Node
+from ._nodered import RED
 from ._property import (
     InputProperty, ListProperty, DictProperty,
     SpinnerProperty, ComboBoxProperty, CodeProperty
 )
-from ._server import Server, StandaloneServer
 
 
-__version__ = "0.1.4"
+__version__ = "0.2.0"
 __author__ = "oyajiDev"
 __email__ = "this.dev.somehit@gmail.com"
 
 __all__ = [
-    "RED", "Node",
+    "RED",
     "InputProperty", "ListProperty", "DictProperty",
     "SpinnerProperty", "ComboBoxProperty", "CodeProperty",
-    "Server", "StandaloneServer"
 ]