+++ noderedpy/_nodered.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
-import os, sys, subprocess, threading, traceback, noderedpy
+import os, sys, subprocess, shutil, json, traceback
 from types import MethodType
 from typing import Type, List
+from glob import glob
 from .templates import package_json, node_html, node_js
 from ._property import Property
 from . import __path__
@@ -11,7 +12,9 @@ class RED:
     """
     Node-RED manager class
     """
-    def __init__(self, user_dir:str, admin_root:str, port:int = 1880, show_default_category:bool = True):
+    registered_nodes:List["Node"] = []
+
+    def __init__(self, user_dir:str, node_red_dir:str = None, admin_root:str = "/node-red-py", port:int = 1880, show_default_category:bool = True, editor_theme:dict = {}):
         """
         Set configs of Node-RED and setup
 
@@ -19,90 +22,204 @@ class RED:
         ----------
         user_dir: str, required
             userDir of Node-RED settings
-        admin_root: str, required
+        node_red_dir: str, default None
+            directory for Node-RED starter
+        admin_root: str, default "node-red-py"
             httpAdminRoot of Node-RED settings
         port: int, default 1880
             port of Node-RED server
         show_default_category: bool, default True
             show default categories of Node-RED or not
+        editor_theme: dict, default {}
+            editorTheme of Node-RED server (for detail information, see https://github.com/node-red/node-red/wiki/Design:-Editor-Themes)
         """
-        self.user_dir, self.admin_root, self.port, self.show_default_category =\
-            user_dir, admin_root, port, show_default_category
+        self.user_dir, self.admin_root, self.port, self.show_default_category, self.editor_theme =\
+            user_dir, admin_root, port, show_default_category, self.__default_editor_theme(editor_theme)
+        
+        # set node_red_dir
+        if node_red_dir is None:
+            self.node_red_dir = os.path.join(__path__[0], "node-red-starter")
+        else:
+            self.node_red_dir = node_red_dir = os.path.realpath(node_red_dir)
+            if os.path.exists(node_red_dir):
+                if not { "index.js", "package.json" }.issubset(set(os.listdir(node_red_dir))):
+                    raise RuntimeError("Target `node_red_dir` is not Node-RED dir format!")
+            else:
+                os.mkdir(node_red_dir)
+                shutil.copyfile(os.path.join(__path__[0], "node-red-starter", "index.js"), os.path.join(node_red_dir, "index.js"))
+                shutil.copyfile(os.path.join(__path__[0], "node-red-starter", "package.json"), os.path.join(node_red_dir, "package.json"))
 
         # setup Node-RED starter
         subprocess.call(
             [ "npm", "install" ],
             stdout = subprocess.DEVNULL,
-            cwd = os.path.join(__path__[0], "node-red-starter")
+            cwd = self.node_red_dir
         )
 
-    # default started callback
-    def __on_started(self, callback:MethodType = None):
+    # create default editor_theme
+    def __default_editor_theme(self, editor_theme:dict):
+        page_theme = editor_theme.pop("page", {})
+        page_theme.update({
+            "title": page_theme.pop("title", "Node-RED.py"),
+            "favicon": page_theme.pop("favicon", os.path.join(__path__[0], "assets", "python-logo.png"))
+        })
+        header_theme = editor_theme.pop("header", {})
+        header_theme.update({
+            "title": "Node-RED.py",
+            "image": None
+        })
+        project_feature = editor_theme.pop("projects", {})
+        project_feature.update({
+            "enabled": project_feature.pop("enabled", False)
+        })
+
+        editor_theme.update({
+            "page": page_theme,
+            "header": header_theme,
+            "userMenu": editor_theme.pop("userMenu", False),
+            "projects": project_feature
+        })
+
+        return editor_theme
+    
+    def register(self, node_func:MethodType, name:str, category:str = "nodered_py", properties:List[Property] = []):
+        """
+        Function to register Node function
+
+        Parameters
+        ----------
+        node_func: MethodType, required
+            Node function to register
+        name: str, required
+            name of Node to register
+        category: str, default nodered_py
+            category of Node
+        properties: List[noderedpy._property.Property]
+            propertis of Node
+        """
+        node = Node(name if name.startswith("nodered-py") else f"nodered-py-{name}", category, properties, node_func)
+        RED.registered_nodes.append(node)
+
+    # check input and run node
+    def __check_input_from_node(self):
+        input_file, output_file = os.path.join(self.__cache_dir, "input.json"), os.path.join(self.__cache_dir, "output.json")
+
         while True:
-            if os.path.exists(self.__started_file):
-                break
+            if os.path.exists(input_file):
+                # read input file
+                while True:
+                    # if cannot read file or read during file writing, read file until can read
+                    try:
+                        with open(input_file, "r", encoding = "utf-8") as ifr:
+                            input_data = json.load(ifr)
+                        
+                        os.remove(input_file)
+                        break
+                    except json.JSONDecodeError:
+                        pass
 
-        callback()
+                node = list(filter(lambda n: n.name == input_data["name"], RED.registered_nodes))[0]
 
-    def start(self, wait:bool = False, debug:bool = True, callback:MethodType = None, server:Type["noderedpy._server.Server"] = None):
+                with open(output_file, "w", encoding = "utf-8") as ofw:
+                    json.dump(node.run(input_data["props"], input_data["msg"]), ofw, indent = 4)
+
+    def start(self, debug:bool = True, callback:MethodType = None):
         """
         Start Node-RED server
 
         Parameters
         ----------
-        wait: bool, default False
-            wait for Node-RED server process of not
         debug: bool, default True
             show outputs on console or not
         callback: MethodType, default None
             callback when Node-RED server started
-        server: Type[noderedpy._server.Server], default None
-            server for setup "user category"
         """
 
+        # setup user_dir
+        self.__start_for_ready()
+
         # set started flag file
-        self.__started_file = os.path.join(__path__[0], "node-red-starter", "started")
+        self.__started_file = os.path.join(self.node_red_dir, "started")
         # kill if process listen on port
         self.stop()
 
-        # map callback
-        if callback:
-            threading.Thread(target = self.__on_started, args = (callback,), daemon = True).start()
+        # set cache_dir
+        self.__cache_dir = os.path.join(self.user_dir, ".cache")
+        if os.path.exists(self.__cache_dir):
+            shutil.rmtree(self.__cache_dir)
 
-        # set subprocess args
-        args = [
+        os.mkdir(self.__cache_dir)
+
+        # save editor_theme
+        with open(os.path.join(self.node_red_dir, "editorTheme.json"), "w", encoding = "utf-8") as tjw:
+            json.dump(self.editor_theme, tjw)
+
+        # remove existing nodes
+        for node_dir in glob(os.path.join(self.user_dir, "node_modules", "nodered-py-*")):
+            shutil.rmtree(node_dir)
+
+        # create custom nodes
+        for node in RED.registered_nodes:
+            node.create(self.user_dir, self.__cache_dir)
+
+        # run Node-RED server
+        subprocess.Popen([
             "node",
             "index.js",
             f"--user-dir={self.user_dir}",
             f"--admin-root={self.admin_root}",
             f"--port={self.port}",
             f"--show-default-category={'true' if self.show_default_category else 'false'}"
-        ]
-        if server:
-            args.append(
-                f"--user-category={','.join(list(set([ node.category for node in server.registered_nodes ])))}"
-            )
+            f"--user-category={','.join(list(set([ node.category for node in RED.registered_nodes ])))}"
+        ], shell = False, stdout = sys.stdout if debug else subprocess.DEVNULL, cwd = self.node_red_dir)
 
-        # run Node-RED server
-        proc = subprocess.Popen(
-            args, shell = False, stdout = sys.stdout if debug else subprocess.DEVNULL,
-            cwd = os.path.join(__path__[0], "node-red-starter")
-        )
-        if wait:# wait if flag is True
-            proc.wait()
+        while True:
+            if os.path.exists(self.__started_file):
+                if callback:
+                    callback()
 
-    def start_for_ready(self):
+                break
+
+        try:
+            self.__check_input_from_node()
+        except KeyboardInterrupt:
+            self.stop()
+
+    def __start_for_ready(self):
         """
         Start Node-RED for setup default userDir
         """
-        def on_started():
+        import time
+
+        self.__started_file = os.path.join(self.node_red_dir, "started")
+        if os.path.exists(self.__started_file):
             os.remove(self.__started_file)
-            self.__stop()
 
-        self.start(True, False, on_started)
+        subprocess.Popen([
+            "node",
+            "index.js",
+            f"--user-dir={self.user_dir}",
+            f"--admin-root={self.admin_root}",
+            f"--port={self.port}",
+            f"--show-default-category={'true' if self.show_default_category else 'false'}"
+            f"--user-category={','.join(list(set([ node.category for node in RED.registered_nodes ])))}"
+        ], shell = False, stdout = subprocess.DEVNULL, cwd = self.node_red_dir)
 
-    def __stop(self):
+        while True:
+            if os.path.exists(self.__started_file):
+                self.stop()
+                break
+
+        time.sleep(1)
+
+    def stop(self):
+        """
+        Stop Node-RED server
+        """
         import psutil, signal
+
+        if os.path.exists(self.__started_file):
+            os.remove(self.__started_file)
 
         killed = False
         for process in psutil.process_iter():
@@ -112,20 +229,11 @@ class RED:
                         process.send_signal(signal.SIGTERM)
                         killed = True
                         break
-            except psutil.AccessDenied:
+            except ( psutil.AccessDenied, psutil.ZombieProcess ):
                 pass
 
             if killed:
                 break
-
-    def stop(self):
-        """
-        Stop Node-RED server
-        """
-        if os.path.exists(self.__started_file):
-            os.remove(self.__started_file)
-
-        self.__stop()
 
 class Node:
     def __init__(self, name:str, category:str, properties:List[Property], node_func:MethodType):
@@ -140,7 +248,7 @@ class Node:
         
         self.__node_func = node_func
 
-    def create(self, node_red_user_dir:str, port:int):
+    def create(self, node_red_user_dir:str, node_red_user_cache_dir:str):
         node_dir = os.path.join(node_red_user_dir, "node_modules", self.name)
         os.makedirs(os.path.join(node_dir, "lib"))
 
@@ -151,10 +259,10 @@ class Node:
             nhw.write(node_html(self))
 
         with open(os.path.join(node_dir, "lib", f"{self.name}.js"), "w", encoding = "utf-8") as njw:
-            njw.write(node_js(self, port))
+            njw.write(node_js(self, node_red_user_cache_dir))
 
     def run(self, props:dict, msg:dict) -> dict:
-        print(f"{self.name} started\n===================================")
+        print(f"\n{self.name} started\n===================================")
         try:
             resp = self.__node_func(props, msg)
             print("============================= ended\n")