+++ noderedpy/decorator.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from typing import List
 from types import MethodType
-from ._nodered import Node
+from ._nodered import RED, Node
 from ._property import Property
 from ._server import Server
 
@@ -21,7 +21,7 @@ def register(name:str, category:str = "nodered_py", properties:List[Property] = 
     """
     def decorator(node_func:MethodType):
         node = Node(name if name.startswith("nodered-py") else f"nodered-py-{name}", category, properties, node_func)
-        Server.registered_nodes.append(node)
+        RED.registered_nodes.append(node)
 
         return node_func